rename anchor card file
